requirements.txt
fastapi
uvicorn
langchain
openai
langchain-openai
python-dotenv
streamlit
requests
langchain_groq